Improve scene

+++ cad_pipeline/cube.h
@@ -14,7 +14,7 @@ AnyGeometry MakeCube(float half_extent) {
   auto cube = pmp::hexahedron();
   
   for (const auto& v : cube.vertices()) {
-    cube.position(v) *= sqrt(3.0) * half_extent;
+    cube.position(v) *= sqrtf(3.f) * half_extent;
   }
 
   pmp::face_normals(cube);
